feat: extract territory pumping utilities to own trait
